+++ DTCD-StyleSystem/src/base-components/BaseInput/BaseInput.js
@@ -32,6 +32,8 @@ export default class BaseInput extends HTMLElement {
       'size',
       'readonly',
       'invalid',
+      'maxlength',
+      'minlength',
     ];
   }
 
@@ -134,6 +136,14 @@ export default class BaseInput extends HTMLElement {
 
       case 'invalid':
         this.invalid = newValue;
+        break;
+
+      case 'maxlength':
+        this.maxlength = newValue;
+        break;
+
+      case 'minlength':
+        this.maxlength = newValue;
         break;
 
       default:
@@ -273,6 +283,30 @@ export default class BaseInput extends HTMLElement {
       this.setAttribute('size', value);
     } else {
       this.removeAttribute('size');
+    }
+  }
+
+  get maxlength() {
+    return this.#internalInput.getAttribute('maxlength');
+  }
+
+  set maxlength(value) {
+    if (value) {     
+      this.#internalInput.setAttribute('maxlength', value);
+    } else {
+      this.#internalInput.removeAttribute('maxlength');
+    }
+  }
+
+  get minlength() {
+    return this.#internalInput.getAttribute('minlength');
+  }
+
+  set minlength(value) {
+    if (value) {     
+      this.#internalInput.setAttribute('minlength', value);
+    } else {
+      this.#internalInput.removeAttribute('minlength');
     }
   }